gas_limit as Weight

+++ runtimes/qf-runtime/src/configs/mod.rs
@@ -305,7 +305,7 @@ parameter_types! {
 	pub const PolkaVmMaxCodeLen: u32 = 131072;
 	pub const PolkaVmMaxCodeVersion: u64 = u64::MAX;
 	pub const PolkaVmMaxUserDataLen: u32 = 2048;
-	pub const PolkaVmMaxGasLimit: u32 = 2097152;
+	pub const PolkaVmMaxGasLimit: u64 = 2097152;
 	pub const PolkaVmMaxStorageKeySize: u32 = 256;
 	pub const PolkaVmMaxStorageSlots: u32 = 4;
 	pub const PolkaVmMinGasPrice: u64 = 1;
